Split test_model.py into two separate tests
